+++ test/6/5.3/run.sh
@@ -10,5 +10,6 @@ fi
 
 docker-compose up -d
 docker-compose exec mariadb make check-ready -f /usr/local/bin/actions.mk
+docker-compose exec php chown -R www-data:www-data .
 docker-compose exec --user=82 php ./test.sh
 docker-compose down